[Feat] 수파베이스 데이터 가져오기
