+++ src/components/myPageComponent/MyPageHeader.jsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const MyPageHeader = () => {
+  return <div>MyPageHeader</div>;
+};
+
+export default MyPageHeader;
